Clean up client secret parsing. Move it into the crypto package instead of ugrpc.
