+++ pkg/crypto/client_secret_test.go
@@ -1,4 +1,4 @@
-package ugrpc
+package crypto
 
 import (
 	"testing"
@@ -6,7 +6,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestParseSecret(t *testing.T) {
+func TestParseClientSecret(t *testing.T) {
 	badSecrets := []string{
 		"v1:1234567890:1234567890:1234567890",
 		"v1:1234567890:1234567890:1234567890:1234567890",
@@ -15,7 +15,7 @@ func TestParseSecret(t *testing.T) {
 	}
 	for _, badSecret := range badSecrets {
 		secretBytes := []byte(badSecret)
-		secretKey, err := ParseSecret(secretBytes)
+		secretKey, err := ParseClientSecret(secretBytes)
 		require.Error(t, err)
 		require.Nil(t, secretKey)
 	}
@@ -24,7 +24,7 @@ func TestParseSecret(t *testing.T) {
 	validSecret := "secret-token:conductorone.com:v1:eyJrdHkiOiJPS1AiLCJjcnYiOiJFZDI1NTE5IiwieCI6InpwM1BXY2NTQlZPdGhTTnN5ck1zVGVFaDlpTjhTWTNnSDZfcDBEUkZaTWMi" +
 		"LCJkIjoielkzTzFybnItQUV3NW0wbVpCTzBjSEZpWWoxTmlSdk1pU1Y1X1Ffa0tiZyJ9"
 	secretBytes := []byte(validSecret)
-	secretKey, err := ParseSecret(secretBytes)
+	secretKey, err := ParseClientSecret(secretBytes)
 	require.NoError(t, err)
 	require.NotNil(t, secretKey)
 }
