+++ pkg/ugrpc/c1_credential_provider.go
@@ -1,10 +1,7 @@
 package ugrpc
 
 import (
-	"bytes"
 	"context"
-	"crypto/ed25519"
-	"encoding/base64"
 	"errors"
 	"net"
 	"net/http"
@@ -13,6 +10,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/conductorone/baton-sdk/pkg/crypto"
 	"github.com/conductorone/baton-sdk/pkg/uhttp"
 	"github.com/go-jose/go-jose/v4"
 	"github.com/go-jose/go-jose/v4/json"
@@ -30,10 +28,7 @@ const (
 )
 
 var (
-	ErrInvalidClientSecret = errors.New("invalid client secret")
-	ErrInvalidClientID     = errors.New("invalid client id")
-
-	v1SecretTokenIdentifier = []byte("v1")
+	ErrInvalidClientID = errors.New("invalid client id")
 )
 
 type c1Token struct {
@@ -64,43 +59,6 @@ func parseClientID(input string) (string, string, error) {
 	}
 
 	return clientName, items[0], nil
-}
-
-func ParseSecret(input []byte) (*jose.JSONWebKey, error) {
-	return parseSecret(input)
-}
-
-func parseSecret(input []byte) (*jose.JSONWebKey, error) {
-	items := bytes.SplitN(input, []byte(":"), 4)
-	if len(items) != 4 {
-		return nil, ErrInvalidClientSecret
-	}
-
-	if !bytes.Equal(items[2], v1SecretTokenIdentifier) {
-		return nil, ErrInvalidClientSecret
-	}
-
-	jwkData, err := base64.RawURLEncoding.DecodeString(string(items[3]))
-	if err != nil {
-		return nil, ErrInvalidClientSecret
-	}
-
-	npk := &jose.JSONWebKey{}
-	err = npk.UnmarshalJSON(jwkData)
-	if err != nil {
-		return nil, ErrInvalidClientSecret
-	}
-
-	if npk.IsPublic() || !npk.Valid() {
-		return nil, ErrInvalidClientSecret
-	}
-
-	_, ok := npk.Key.(ed25519.PrivateKey)
-	if !ok {
-		return nil, ErrInvalidClientSecret
-	}
-
-	return npk, nil
 }
 
 func (c *c1TokenSource) Token() (*oauth2.Token, error) {
@@ -201,7 +159,7 @@ func newC1TokenSource(ctx context.Context, clientID string, clientSecret string)
 		return nil, "", "", err
 	}
 
-	secret, err := parseSecret([]byte(clientSecret))
+	secret, err := crypto.ParseClientSecret([]byte(clientSecret))
 	if err != nil {
 		return nil, "", "", err
 	}
